Untangle ID generation race condition in constant registration
The nested call to `host.to_polar(cls)` within the `register_constant` function
causes a race condition with respect to the generation of the ID under which the
class will be registered. Specifically `host.to_polar(v)` ends up calling
`cache_instance` first with a `None` ID *right before* `register_class` attempts
to do the same with its own `ID` value.

The (hacky) solution is to extend `to_polar` to optionally take an existing
`instance_id` which it will use in lieu of generating a new one. In addition
the FFI `register_constant` API is updated to *receive* an ID value rather than
generate and return one.